Elizabeth

+++ LearnLynks/src/main/java/com/project/LearnLynks/dtos/response/ArchiveCurriculumResponse.java
@@ -1,15 +1,15 @@
 package com.project.LearnLynks.dtos.response;
 
 public class ArchiveCurriculumResponse {
-    public int getCurriculumId() {
+    public Long getCurriculumId() {
         return curriculumId;
     }
 
-    public void setCurriculumId(int curriculumId) {
+    public void setCurriculumId(Long curriculumId) {
         this.curriculumId = curriculumId;
     }
 
-    private int curriculumId;
+    private Long curriculumId;
     private String message;
 
     public String getMessage() {
